HOCS-2457 : add separate swagger config file
