+++ src/main/resources/application.properties
@@ -58,3 +58,4 @@ aws.sqs.access.key=12345
 audit.aws.sqs.access.key=12345
 
 user.email.whitelist=homeoffice.gov.uk,hmpo.gov.uk
+
